requirements.txt
torch==2.11.0
yapper-tts
soundfile
scipy
numpy

